requirements.txt
altgraph==0.17.4
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
idna==3.10
importlib-metadata==6.7.0
packaging==24.0
Pillow==9.5.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.7
PyJWT==2.8.0
python-dotenv==0.21.1
pytz==2024.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
whitenoise==6.5.0
zipp==3.15.0
